Finished first autonomous movement build
Just finished making the first version of our timed movement code. Going to work on making it better for a series of these commands, as I'm not sure if it will wait for the while statements in each instance to finish before moving on to the next, or if it will try to call all instances at once.